Updates post-debian8

+++ runDebarcer.sh
@@ -22,6 +22,7 @@ LAUNCHDIR=$PWD
 OUTPUTDIR=$LAUNCHDIR
 DBROOT=$BHOME
 DBSRC=$DBROOT"/src/"
+USE_SGE=false
 
 usage()
 {
@@ -86,11 +87,13 @@ if [[ $ONLYGRAPHICS ]]; then
 fi
 
 
-VERSIONID="0.3.1"
+VERSIONID="0.4.0"
 STARTTIME=$(date +"%Y%m%d-%H%M%S")
 MAINLOG="DeBarcEr."$STARTTIME".log"
+PERLVERSION=`perl -e 'print $^V';`
 echo "[Debarcer `date`] : Logfile timestamp: $STARTTIME" > $MAINLOG
 echo "Running Debarcer version $VERSIONID" >> $MAINLOG
+echo "Running perl version $PERLVERSION" >> $MAINLOG
 echo "---" >> $MAINLOG
 echo "git info:" >> $MAINLOG
 cat $DBROOT/.git/HEAD >> $MAINLOG
@@ -118,7 +121,9 @@ fi
 echo "[Debarcer `date`] Running workflow from $BHOME" >> $MAINLOG
 
 CONFIG_FILE="$DBROOT/config/debarcer.conf"
-### FIXME Need to dump the config file to the log here ###
+echo "Using debarcer config file:" >> $MAINLOG
+cat $CONFIG_FILE >> $MAINLOG;
+echo "End of debarcer config file." >> $MAINLOG
 
 # Some bwa mem files for future development
 #
@@ -149,31 +154,44 @@ rm Rplots.pdf # Remove empty plot just made by ggplot
 
 ## FIXME Since the UID depths file is available the composition can be analyzed here using src/_calculateBarcodeBiases.pl
 
-echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=1" >> $MAINLOG
-# Arguments: --sampleID; --consDepth; --plexity ... others.
-qsub -N DbC1$SAMPLENAME -l h_vmem=16G -e DbC1$SAMPLENAME.log -o DbC1$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=1 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons1.txt"
-	
-echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=3" >> $MAINLOG
-# Arguments: --sampleID; --consDepth; --plexity ... others.
-qsub -N DbC3$SAMPLENAME -l h_vmem=16G -e DbC3$SAMPLENAME.log -o DbC3$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=3 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons3.txt"
-	
-echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=10" >> $MAINLOG
-# Arguments: --sampleID; --consDepth; --plexity ... others.
-qsub -N DbC10$SAMPLENAME -l h_vmem=16G -e DbC10$SAMPLENAME.log -o DbC10$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=10 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons10.txt"
-	
-echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=20" >> $MAINLOG
-# Arguments: --sampleID; --consDepth; --plexity ... others.
-qsub -N DbC20$SAMPLENAME -l h_vmem=16G -e DbC20$SAMPLENAME.log -o DbC20$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=20 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons20.txt"
-	
-echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=30" >> $MAINLOG
-# Arguments: --sampleID; --consDepth; --plexity ... others.
-qsub -N DbC30$SAMPLENAME -l h_vmem=16G -e DbC30$SAMPLENAME.log -o DbC30$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=30 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons30.txt"
 
-qsub -N "AggregateDebarcers" -hold_jid DbC1$SAMPLENAME,DbC3$SAMPLENAME,DbC10$SAMPLENAME,DbC20$SAMPLENAME,DbC30$SAMPLENAME -sync y -cwd -b y -e ./sge -o ./sge "sleep 1"
+if [ $USE_SGE = true ]; then
 
-cat DbC*.log >> $MAINLOG  # Aggregate the log files generated above.
-rm DbC*.log
+	echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=1" >> $MAINLOG
+	# Arguments: --sampleID; --consDepth; --plexity ... others.
+	qsub -N DbC1$SAMPLENAME -l h_vmem=16G -e DbC1$SAMPLENAME.log -o DbC1$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=1 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons1.txt"
+		
+	echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=3" >> $MAINLOG
+	# Arguments: --sampleID; --consDepth; --plexity ... others.
+	qsub -N DbC3$SAMPLENAME -l h_vmem=16G -e DbC3$SAMPLENAME.log -o DbC3$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=3 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons3.txt"
+		
+	echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=10" >> $MAINLOG
+	# Arguments: --sampleID; --consDepth; --plexity ... others.
+	qsub -N DbC10$SAMPLENAME -l h_vmem=16G -e DbC10$SAMPLENAME.log -o DbC10$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=10 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons10.txt"
+		
+	echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=20" >> $MAINLOG
+	# Arguments: --sampleID; --consDepth; --plexity ... others.
+	qsub -N DbC20$SAMPLENAME -l h_vmem=16G -e DbC20$SAMPLENAME.log -o DbC20$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=20 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons20.txt"
+		
+	echo "[Debarcer `date`] BAM Consensus for $SAMPLENAME depth	=30" >> $MAINLOG
+	# Arguments: --sampleID; --consDepth; --plexity ... others.
+	qsub -N DbC30$SAMPLENAME -l h_vmem=16G -e DbC30$SAMPLENAME.log -o DbC30$SAMPLENAME.log -cwd -b y "module load debarcer; time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=30 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons30.txt"
 
+	qsub -N "AggregateDebarcers" -hold_jid DbC1$SAMPLENAME,DbC3$SAMPLENAME,DbC10$SAMPLENAME,DbC20$SAMPLENAME,DbC30$SAMPLENAME -sync y -cwd -b y -e ./sge -o ./sge "sleep 1"
+
+	cat DbC*.log >> $MAINLOG  # Aggregate the log files generated above.
+	rm DbC*.log
+
+else
+
+	# Do the commands serially on the local node
+	time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=1 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons1.txt
+	time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=3 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons3.txt
+	time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=10 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons10.txt
+	time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=20 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons20.txt
+	time perl $DBSRC/generateConsensusFromBAM.pl --bam=$SAMPLEPREFIX.sorted.bam --sampleID=$SAMPLENAME --consDepth=30 --config=$CONFIG_FILE > ./tables/$SAMPLENAME.bamPositionComposition.cons30.txt
+
+fi
 
 # FIXME generate nucleotide context error tables
 # Still in development
